renamed table_styles.py to figure_styles.py
