+++ app.py
@@ -10,7 +10,7 @@ import pandas as pd
 import sys
 from datetime import datetime
 from nba_data import scatter_data, conf_table_data, team_colors, scatter_vals, conf_table_cols
-from table_styles import conf_table_params, center_style
+from figure_styles import conf_table_params, center_style
 
 
 playoff_splitter = lambda x: [j.rsplit(' -')[0] for j in x['Team'].values]
